[web] Add sources/gauges status indicator

+++ build/scripts/dev_cleanup.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
 
 docker-sync-stack clean
-docker rm --force --volumes $(docker ps -a --filter="label=guide.whitewater" -q)
-docker image rm $(docker image ls --filter "label=guide.whitewater")
-docker system prune --force --volumes
+
+if [[ $* == *--hard* ]]; then
+    docker rm --force --volumes $(docker ps -a --filter="label=guide.whitewater" -q)
+    docker image rm $(docker image ls --filter "label=guide.whitewater")
+    docker system prune --force --volumes
+fi
